* server/character.h, server/movie_def_impl.cpp,           server/movie_def_impl.h, server/movie_definition.h,           server/movie_interface.h, server/movie_root.cpp,           server/movie_root.h, server/sprite_definition.h,           server/sprite_instance.cpp, server/sprite_instance.h:           Fixed more frame number types to be size_t rather then int;           added some doxygen comments; fixed ensure_frame_loaded calls           to not erroneusly using sprite frames rather then movie frames;           removed ensure_frame_loaded calls from goto_frame (will need           further thinking about it).

+++ server/character.h
@@ -207,7 +207,7 @@ public:
 		return get_relative_target_common(name);
 	}
 
-    virtual int	get_current_frame() const { assert(0); return 0; }
+    virtual size_t	get_current_frame() const { assert(0); return 0; }
     virtual bool	has_looped() const { assert(0); return false; }
     virtual void	restart() { /*assert(0);*/ }
